NROM: Fixed Roms getting cut off
Previously the functionality wrote the rom file in two chunks, one for
each bank. But now it writes it in one go. Though a leftover from the
previous set up was that the data array was only the length required for
the first chunk write.

Now we have access to the full rom to do a full write.