BreakpointInputBox: Result is only checked when OK was hit
If the user closes the form by hitting ok, cancel, or the X button, a
function is called to catch that.

The purpose is to check the validate the input before accepting the OK
button. If valid, it would allow the form to close, otherwise it will
cancel.

However, how it was set up it would do this even if one pressed cancel.
Now it will only happen if the user has pressed OK.