CPU: Fixed ADC Carry Flag Misdetections
A mistake was made  in which the sum included the Carry_Flag itself
instead of the variable which contains 1 or 0 depending on whether the
carry flag is set.

This is not correct, because Carry_Flag is a constant value used to set
and mask the carry flag in the processor status register. Instead, we
want to use Carry_Flag to find out if the carry flag is actually set and
add that into the sum.

This is now fixed.