CPU: Fixed ADC sum calculation issue
Previously the sum was being casted to a byte. While this should be done
to set the accumulator, the sum in integer form is still required to
accurately set the carry flag.

Now the sum result is stored as an integer and is only casted during the
setting of the accumulator.