PPU: Added function that can draw line from individual nametable tile
Previously, functions were added to draw a full 512x1 line then select
the 256x1 pixels needed for the frame. The reason for this was it made
it easy to account for side by side nametables, and scrolling across
them.

The problem with this was that it was a lot more work. There was a lot
of data being drawn that would not be used in the frame. This caused
slow-downs due to the increased work.

Now, this function allows drawing of lines on a tile by tile basis. This
will be important as it will allow drawing only the tiles required for
the frame.