PPU: Added function that can draw 256x1 frame line from only tiles required
This function makes use of the new tile by tile drawing function.

Instead of drawing a full 512x1 line and selecting the 256 being used,
it draws the tiles needed for the 256 line and constructs the line from
the pixels required of those tiles.

This requires less tiles to draw per line and an overall decrease of
work per line.