PPU: Removed function that draws frame line from 512x512 window
This function required an entire 512x1 line to be drawn in order to draw
the 256x1 line needed for the frame. The reason for this was so that
both the left and right name tables for that line would be drawn. Then
this function could easily scroll across that 512x1 line without any
trouble and draw the frame line with ease.

The downside was that because it had to draw the full 512x1 line it was
quite inefficient. It drew many tiles that would not get used for the
frame at all. And since this needs to be done 240 times per frame, it
adds up to be quite slow, despite it's advanges over the original BG
line drawing function.

This has since been antiquated. A new function has been added that draws
the line tile by tile and only draws the tiles needed.