MMC1: Fixed error in 8kb chr bank swaps
A mistake was made in calculating the bankAddress for the bank to read.
Although in 8kb banking mode you can only access half as many banks, the
size of the banks are still the same.

Bank 0 is still 0x0000
Bank 1 is still 0x1000
Bank 2 is still 0x2000

But in 8b mode, 0 accesses 0 and 1, and 2 accesses 2 and 3.
So bank address is calculated the same way regardless of bank size. It
only effects which banks are allowed to be accessed.