PPU: Fixed Tile Bugginess when scrolling vertically
Previously there were issues when vertically scrolling which included
tiles getting the wrong attributes on them, or drawing garbled graphics
on two y tile positions. This was due to not accounting for the fact
that the screen is not a square. It is 32x30 tiles, but the modulus
operation for x and y tile positions were both accounting for 32 tiles.
The y tile was meant to use 30.