PPU: Retrieval of sprite pattern table address now accounts for 8x16 pixel sprites
Previously the tileID was used directly to index the pattern table in
all cases. 8x16 sprites don't do this precisely. For 8x16 sprites, bit 0
of the tile ID selects the pattern table address (as opposed to using
PPUCTRL register to decide. Then the tileID with bit 0 masked off
becomes the tileID to index with.)