MainWindow: Clean-up/Fixes
An issue was created through committing sections of files at a time where semi-colons that are not in the source file are still left in the GitHub source.

This went unnoticed due to it not existing in the local repository and was caught on attempt to clone and build the repository on another machine.