Configured DLLMaps to point SFMLNet DLLs to installed CSFML linux .so files
Previously if one attempted to build and run NEST in Mono on Linux they would get errors on runtime due to the SFMLNet DLLs not being able to find "csfml-graphics-2" or the audio, network, and window equivalents.

The installed files for 2.4 are not named this and therefore would not be found. Therefore a DLL Map has been configured so that the SFMLNet dlls will instead search for libcsfml-graphics.so.2.4 instead and the other equivalents.

If installed properly the .so files should be found, allowing NEST to run on Linux.