Replaced fetchAndExecute if statements with switch
This was originally done so that the series of if statements could be easily generated, reducing the likelyness of a typographic mistake.

Then, the opcodes that did not exist could simply be removed one by one.

This was effective in reducing issues, however, it is possible to easily convert these ifs into a switch statement now that there is no need to modify these.

This should allow the switch to be optimized into a lookup table.