Updated readme for name change
Now that NEST has a C++ sister project, NEST has been renamed to NEST-Sharp with the NEST namesake being given to the C++ cross platform version. The readme has been updated with the new name, a reference to the C++ version, and changed the image URLs to make the previews display post name change.