Can now properly load roms consecutively
Previously if one loaded a rom file to run, it would work as normal. However when they then opted to load another rom to emulate instead in the same run time, the next one would likely not operate correctly like it would if it was the first rom file to be loaded in.

This was due to not resetting the state of the emulation classes when loading in the second rom file. If the cpu was in the middle of an interrupt, or the GPU was in a specific render state, these would still be the case when the new rom started. This could significantly break the emulation of the new rom as it is not intended to be in those states initially.

This has now been fixed by adding functions to reset each class to their initial state and calling that when ready to load a new rom in. This allows each rom to be loaded correctly without problems.