Fixed building MainWindow in MonoDevelop
The UI for GameBeak-Sharp was built in WinForms. When building in Visual Studio an issue never went noticed, which was the saveToolStripMenuItem not being initialized properly in the Designer.cs class.

However, MonoDevelop is strict and requires this to be done manually. This has been fixed, allowing the save toolstrip to build properly with Mono. This will allow it to run on other platforms.