Core reset will no longer occur on first emulation run
The core reset functionality that resets the emulation class state when loading a rom was only intended to occur if emulation of a rom had occurred previously during the application's life time.

This is because the variables would be at the same state as the reset functionality would set it to, therefore does not require the extra work. However, it occurred regardless due to an oversight.

the Core.run variable now instantiates to false and is only set true when emulation is ready. It will be set true right before emulation begins, and the next tiem a rom is loaded, it will check if run is already true. If it is, that means a rom has been loaded prior, and that the state should then be reset.