Update about Archival.
This repository will be archived and will no longer be updated in tandem with the C++ version. In order to get the interface to truly work properly cross platform, it would be best to simply rewrite the GUI for the project. However, the advantages to this would not be clear. The Qt version already works cross platform and only recently been completely moved to Qt. The C# version would likely still not perform as well and would be losing it's primary advantage over the C++ version which is the debugger that has been made for the C# version.

While this project has taken several forms and been rewritten multiple times, each time has had a purpose. However, future work on this version would be additional work for a likely inferior experience. Therefore development will now be moved to focus on a single version. Thanks for the interest in this project.